+++ mesh/mesh/common/TrackBall.h
@@ -10,11 +10,11 @@ namespace hj
   public:
 
     // map 2D screen point to unit sphere
-    static glm::dvec3 MapToSphere(const glm::dvec2& point2d, const glm::ivec2& view_size);
+    static glm::vec3 MapToSphere(const glm::vec2& point2d, const glm::ivec2& view_size);
 
     // get rotate matrix.
-    static glm::dmat4 RotateMatrix(const glm::dvec2& old_point2d,
-      const glm::dvec2& new_point2d,
+    static glm::mat4 RotateMatrix(const glm::vec2& old_point2d,
+      const glm::vec2& new_point2d,
       const glm::ivec2& view_size);
   };
 }
